Refactor imports to remove a testing import from standard code
